Remove dotnet WebRTC implementation
Internal issues with this implementation make it difficult to get it to talk nicely with the other WebRTC packages. Working around these issues has required a lot of effort to-date. The tradeoff no longer seems worth it as we've moved our main platforms onto more robust/complete implementations. If there's demand we could look at using the mixed reality package